Omaha #3410 changed lon/lat fields to floats in sfc and air obs location

Former-commit-id: 74a6c3dd20898e9bd4484376ea4e4e471bdec9af [formerly aa991c922354f0a67fac08fee456ca60a30de968 [formerly cdb0a72fcb2345564f2b944fdecd7869fd498767]]
Former-commit-id: aa991c922354f0a67fac08fee456ca60a30de968
Former-commit-id: a7d17bd490f2f0931f3c22ee5484214041a34bd6

+++ edexOsgi/com.raytheon.uf.edex.plugin.qc/src/com/raytheon/uf/edex/plugin/qc/QCScanner.java
@@ -60,6 +60,7 @@ import com.raytheon.uf.edex.plugin.qc.dao.QCDao;
  * May 16, 2013 1869       bsteffen    Remove DataURI column from qc.
  * Aug 30, 2013 2298       rjpeter     Make getPluginName abstract
  * Feb 20, 2014 DR 17098   D. Friedman Filter out invalid lat/lon values.
+ * Jul 23, 2014 3410       bclement    location changed to floats
  * 
  * </pre>
  * 
@@ -208,8 +209,8 @@ public class QCScanner {
                     while (ri < records.length) {
                         QCRecord r = new QCRecord();
                         double obsTime = dObsTime.getDouble(ri);
-                        double lat = dLat.getDouble(ri);
-                        double lon = dLon.getDouble(ri);
+                        float lat = dLat.getFloat(ri);
+                        float lon = dLon.getFloat(ri);
                         if ((obsTime != vObsTimeFillValue)
                                 && ((vObsTimeMissingValue == null) || (vObsTimeMissingValue != obsTime))
                                 && Math.abs(lon) <= 180 && Math.abs(lat) <= 90) {
